\name{myplot.imagematrix}
\alias{myplot.imagematrix}
\title{Plotting an imagematrix object}
\description{
  This function outputs an imagematrix object as an image.
}
\usage{
  myplot.imagematrix(x,y,z, ...)}
\arguments{
  \item{x}{optional x coordinates}
  \item{y}{optional y coordinates}
  \item{z}{target image}
  \item{...}{plotting options}
}
\examples{
   \dontrun{
     data(logo)
     plot(z=logo, main="plot(logo)")
  }
}
\seealso{}
\keyword{}


