\name{mypolygon}
\alias{mypolygon}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ simple wrapper function to plot colored polygons }
\description{
  same as \link{polygon}, execept the value for color is taken from the 1st element of the exra column 'col'
}
\usage{
mypolygon(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ matrix containing columns X,Y,col }
  \item{\dots}{ extra arguments passed to  \link{polygon} }
}
\details{
  
}
\value{
  }
\references{  }
\author{Markus Loecher, Sense Networks <markus@sensenetworks.com>}
\note{ 
}
\seealso{  }
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
