\name{plotOSMtile}

\alias{plotOSMtile}

\title{plots a single OSM tile}

\description{Adds tile to plot}

\usage{plotOSMtile(osmtile, zoom, add = TRUE, raster = TRUE, 

    verbose = 0, ...)}

\arguments{

  \item{osmtile}{tile object}

  \item{zoom}{zoom level}

  \item{add}{Boolean, whether to add to existing plot}

  \item{raster}{Boolean, whether to load rster image}

  \item{verbose}{level of verbosity}

  \item{\dots}{further arguments to be passed to \code{rasterImage}}

}



\value{returns nothing}



\author{Markus Loecher}











