\name{PlotPolysOnStaticMap}
\alias{PlotPolysOnStaticMap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ plots polygons on map}
\description{
  This function plots polygons on a map.
}
\usage{
PlotPolysOnStaticMap(MyMap, polys, col, verbose = 1, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MyMap}{ map image returned from e.g. \code{GetMap()} }
  \item{polys}{ polygons to overlay }
  \item{col}{ (optional) vector of colors, one for each polygon}
  \item{verbose}{ level of verbosity }
  \item{\dots}{further arguments passed to \code{PlotOnStaticMap}}
}
\details{
  }
\value{
  }
\references{ }
\author{ Markus Loecher, Sense Networks <markus@sensenetworks.com> }
\note{ }
\seealso{ \link{PlotOnStaticMap} \link{mypolygon} }
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

