\name{ReadMapTile}
\alias{ReadMapTile}
\title{Reading a map tile object}
\description{
  This function reads a map tile and its associated metaInfo file as an image.
}
\usage{
  ReadMapTile(destfile, METADATA = TRUE)}
\arguments{
  \item{destfile}{image file to read from}
  \item{METADATA}{read MetaInfo as well ?}
}
\examples{
 }
\seealso{}



