\name{mypolygon}
\alias{mypolygon}
\title{simple wrapper function to plot colored polygons}
\description{same as \link{polygon}, execept the value for color is taken from the 1st element of the exra column 'col'}
\usage{mypolygon(x, ...)}
\arguments{
  \item{x}{matrix containing columns X,Y,col}
  \item{\dots}{extra arguments passed to  \link{polygon}}
}



\author{Markus Loecher}





