\name{plotOSMtile}
\alias{plotOSMtile}
\title{plots a single OSM tile}
\description{Adds tile to plot}
\usage{plotOSMtile(osmtile, zoom, add = TRUE, raster = TRUE, 
    verbose = 0, ...)}
\arguments{
  \item{osmtile}{tile object}
  \item{zoom}{zoom level}
  \item{add}{
}
  \item{raster}{
}
  \item{verbose}{level of verbosity}
  \item{\dots}{further arguments to be passed to \code{rasterImage}}
}

\value{returns nothing}

\author{Markus Loecher}





