\name{RgoogleMaps-package}
\alias{RgoogleMaps-package}
\docType{package}
\title{Overlays on Google map tiles in R}
\description{This package serves two purposes: (i) Provide a
comfortable R interface to query the Google server for static
maps, and (ii) Use the map as a background image to overlay
plots within R. This requires proper coordinate scaling.}
\details{
\tabular{ll}{Package: \tab RgoogleMaps\cr
Type: \tab Package\cr
Title: \tab Overlays on Google map tiles in R\cr
Version: \tab 1.2.0.3\cr
Date: \tab 2013-03-15\cr
Depends: \tab R (>= 2.10), graphics, stats, utils, png\cr
Suggests: \tab PBSmapping, maptools, RColorBrewer, sp\cr
Author: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>\cr
Maintainer: \tab "Markus Loecher" <markus.loecher@gmail.com>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Packaged: \tab 2013-03-21 08:39:58 UTC; mloecher\cr
Repository: \tab CRAN\cr
Date/Publication: \tab 2011-08-02 10:32:13\cr}
}
\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>}

\keyword{ package }


