\name{RgoogleMaps-package}

\alias{RgoogleMaps-package}

\alias{RgoogleMaps}

\docType{package}

\title{Overlays on Google map tiles in R}

\description{This package serves two purposes: (i) Provide a

comfortable R interface to query the Google server for static

maps, and (ii) Use the map as a background image to overlay

plots within R. This requires proper coordinate scaling.}

\details{

\tabular{ll}{Package: \tab RgoogleMaps\cr

Type: \tab Package\cr

Title: \tab Overlays on Google map tiles in R\cr

Version: \tab 1.4.1\cr

Date: \tab 2016-09-06\cr

Depends: \tab R (>= 2.10)\cr

Imports: \tab graphics, stats, utils, png, RJSONIO\cr

Suggests: \tab PBSmapping, maptools, sp, loa, RColorBrewer\cr

Author: \tab Markus Loecher\cr

Maintainer: \tab Markus Loecher <markus.loecher@gmail.com>\cr

License: \tab GPL\cr

LazyLoad: \tab yes\cr}

}

\author{Markus Loecher}



\keyword{ package }





